working collection + view details

+++ MyPokemonCollection/src/main/java/app/stats_calculators/IStatsCalculator.java
@@ -5,6 +5,6 @@ import java.util.List;
 import app.data_objects.Nature;
 
 public interface IStatsCalculator {
-    List<Integer> calculateStats(List<Integer> base, List<Integer> ivs,
-                                 List<Integer> evs, int level, Nature nature);
+    List<Long> calculateStats(List<Long> base, List<Long> ivs,
+                                 List<Long> evs, long level, Nature nature);
 }
